refactor folder remove tailwind variable
