+++ src/app/(auth)/login/LoginForm.tsx
@@ -1,0 +1,104 @@
+"use client";
+import { signIn } from "next-auth/react";
+import { useRouter } from "next/navigation";
+import { useState } from "react";
+import {GoogleLoginBtn} from "@/app/components/GoogleLoginBtn";
+
+
+export const LoginForm = () => {
+
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
+  const router = useRouter();
+
+  return (
+    <div className="mt-10 sm:mx-auto sm:w-full sm:max-w-sm ">
+    <div className="space-y-6">
+      <div>
+        <label
+          htmlFor="email"
+          className="block text-sm font-medium leading-6 text-black"
+        >
+          Email address
+        </label>
+        <div className="mt-2">
+          <input
+            id="email"
+            name="email"
+            type="email"
+            autoComplete="email"
+            onChange={(e) => setEmail(e.target.value)}
+            required
+            className="block w-full rounded-md border bg-white/5 py-1.5 text-black shadow-sm ring-1 ring-inset ring-white/10 focus:outline-1 sm:text-sm sm:leading-6 "
+          />
+        </div>
+      </div>
+
+      <div>
+        <div className="flex items-center justify-between">
+          <label
+            htmlFor="password"
+            className="block text-sm font-medium leading-6 text-black"
+          >
+            Password
+          </label>
+          <div className="text-sm">
+            <div
+              onClick={() => router.push("/forgot-password")}
+              className="cursor-pointer font-semibold text-indigo-400 hover:text-indigo-300"
+            >
+              Forgot password?
+            </div>
+          </div>
+        </div>
+
+        <div className="mt-2">
+          <input
+            id="password"
+            name="password"
+            type="password"
+            autoComplete="current-password"
+            onChange={(e) => setPassword(e.target.value)}
+            required
+            className="block w-full rounded-md border bg-white/5 py-1.5 text-black shadow-sm ring-1 ring-inset ring-white/10 focus:outline-1 sm:text-sm sm:leading-6"
+          />
+        </div>
+      </div>
+
+      <div>
+        <button
+          onClick={() =>
+            signIn("credentials", {
+              email,
+              password,
+              redirect: true,
+              callbackUrl: "/",
+            })
+          }
+          disabled={!email || !password}
+          className="disabled:opacity-40 flex w-full justify-center rounded-md bg-indigo-500 px-3 py-1.5 text-sm font-semibold leading-6 text-black shadow-sm hover:bg-indigo-400 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-500 bg-orangeSecondary"
+        >
+          Login
+        </button>
+      </div>
+
+      <div className="inline-flex items-center justify-center w-full">
+        <hr className="w-64 h-px my-8 bg-gray-400 border-0 dark:bg-gray-700"></hr>
+        <span className="absolute px-3 font-medium text-gray-900 -translate-x-1/2 g-gray-400 left-1/2">or</span>
+      </div>
+      <GoogleLoginBtn/>
+    </div>
+    <p className="mt-10 text-center text-sm text-gray-400">
+      Not a member?{" "}
+      <button
+        onClick={() => router.push("register")}
+        className="font-semibold leading-6 text-indigo-400 hover:text-indigo-300"
+      >
+        Register
+      </button>
+    </p>
+  </div>
+  )
+}
+
+export default LoginForm
