+++ src/app/(auth)/login/page.tsx
@@ -1,0 +1,18 @@
+import {LoginForm} from "./LoginForm"
+const Login = () => {
+ 
+  return (
+    <>
+      <div className="flex min-h-full flex-1 flex-col justify-center px-6 py-12 lg:px-8">
+        <div className="sm:mx-auto sm:w-full sm:max-w-sm">
+          <h2 className="mt-10 text-center text-2xl font-bold leading-9 tracking-tight text-black">
+            Login to your account
+          </h2>
+        </div>
+        <LoginForm/>
+      </div>
+    </>
+  );
+};
+
+export default Login;
