+++ src/app/(auth)/register/page.tsx
@@ -1,0 +1,5 @@
+const Register = () => {
+  return <div>Register Up page</div>;
+};
+
+export default Register;
