+++ src/app/(protected)/page.tsx
@@ -7,7 +7,7 @@ export default async function Page() {
   const session = await getServerSession(authOptions);
 
   if (!session) {
-    redirect("/signIn");
+    redirect("/login");
   }
 
   return (
