+++ src/app/components/GoogleLoginBtn.tsx
@@ -1,0 +1,13 @@
+
+
+export const GoogleLoginBtn = () => {
+  return (
+    <button
+    className="font-semibold leading-6 text-indigo-400 hover:text-indigo-300"
+  >
+    Login with Google account
+  </button>
+  )
+}
+
+export default GoogleLoginBtn
