--- src/app/signUp/page.tsx
@@ -1,5 +1,0 @@
-const signUp = () => {
-  return <div>Sign Up page</div>;
-};
-
-export default signUp;
