+++ tailwind.config.ts
@@ -7,15 +7,6 @@ const config: Config = {
     "./src/app/**/*.{js,ts,jsx,tsx,mdx}",
   ],
   theme: {
-    colors: {
-      transparent: "transparent",
-      current: "currentColor",
-      white: "#F7F7F7",
-      black: "#272D2F",
-      orangePrimary: "#FE724C",
-      orangeSecondary: "#CF4307",
-      orangeAccent: "#E78122",
-    },
     extend: {
       backgroundImage: {
         "gradient-radial": "radial-gradient(var(--tw-gradient-stops))",
